requirements.txt
hikari
hikari[speedups]
hikari-lightbulb
python-dotenv
aiosqlite

uvloop
